+++ xblock/test/test_fields.py
@@ -2,6 +2,7 @@
 Tests for classes extending Field.
 """
 
+from __future__ import unicode_literals
 # Allow accessing protected members for testing purposes
 # pylint: disable=W0212
 
@@ -721,5 +722,5 @@ class FieldSerializationTest(unittest.TestCase):
         ['{"foo":"bar"}', '[1, 2, 3]', 'baz', '1.abc', 'defg']))
     def test_from_string_errors(self, _type, string):
         """ Cases that raises various exceptions."""
-        with self.assertRaises(StandardError):
+        with self.assertRaises(Exception):
             _type(enforce_type=True).from_string(string)