+++ xblock/test/test_fields_api.py
@@ -24,8 +24,11 @@ tests of the various operations), and test setup (which set up the
 particular combination of initial conditions that we want to test)
 """
 
+from __future__ import unicode_literals
 import copy
 from mock import Mock
+import six
+from six.moves import range  # pylint: disable=redefined-builtin, import-error
 
 from xblock.core import XBlock
 from xblock.fields import Integer, List, String, ScopeIds, UNIQUE_ID
@@ -424,7 +427,7 @@ class TestImmutableWithComputedDefault(ImmutableTestCases, ComputedDefaultTestCa
 
     @property
     def default_iterator(self):
-        return iter(xrange(1000))
+        return iter(range(1000))
 
 
 class TestMutableWithStaticDefault(MutableTestCases, StaticDefaultTestCases, DefaultValueMutationProperties):
@@ -441,7 +444,7 @@ class TestMutableWithComputedDefault(MutableTestCases, ComputedDefaultTestCases,
 
     @property
     def default_iterator(self):
-        return ([None] * i for i in xrange(1000))
+        return ([None] * i for i in range(1000))
 
 
 class TestImmutableWithInitialValue(ImmutableTestCases, InitialValueProperties):
@@ -504,17 +507,24 @@ class SaveNoopPrefix(object):
 
 for operation_backend in (BlockFirstOperations, FieldFirstOperations):
     for noop_prefix in (None, GetNoopPrefix, GetSaveNoopPrefix, SaveNoopPrefix):
+        # pylint: disable=bad-continuation
         for base_test_case in (
-                TestImmutableWithComputedDefault, TestImmutableWithInitialValue, TestImmutableWithStaticDefault,
-                TestMutableWithComputedDefault, TestMutableWithInitialValue, TestMutableWithStaticDefault,
+                TestImmutableWithComputedDefault, TestImmutableWithInitialValue,
+                TestImmutableWithStaticDefault, TestMutableWithComputedDefault,
+                TestMutableWithInitialValue, TestMutableWithStaticDefault,
                 TestImmutableWithUniqueIdDefault, TestImmutableWithInitialValueAndUniqueIdDefault
         ):
+            # pylint: enable=bad-continuation
 
             test_name = base_test_case.__name__ + "With" + operation_backend.__name__
             test_classes = (operation_backend, base_test_case)
             if noop_prefix is not None:
                 test_name += "And" + noop_prefix.__name__
                 test_classes = (noop_prefix, ) + test_classes
+
+            if six.PY2:
+                # Python 2.x can't handle unicode as an argument to the type() function
+                test_name = test_name.encode('utf-8')
 
             vars()[test_name] = type(test_name, test_classes, {'__test__': True})