+++ xblock/test/test_json_conversion.py
@@ -2,6 +2,7 @@
 Tests asserting that ModelTypes convert to and from json when working
 with ModelDatas
 """
+from __future__ import unicode_literals
 # Allow inspection of private class members
 # pylint: disable=W0212
 from mock import Mock