+++ xblock/test/test_mixins.py
@@ -1,6 +1,7 @@
 """
 Tests of the XBlock-family functionality mixins
 """
+from __future__ import unicode_literals
 
 from unittest import TestCase